Full page rendering complete!
Well, kinda. There is little error checking, and I am having trouble
with some of the plugins to the templates. Not sure why entirely. But
the logic for crawling through the markdown files and rendering to
static text is complete.

Now, we just need error checking, tests, features, etc. etc.